renamed memory allocators for clarity
