+++ security-monitor/src/core/control_data/hardware_hart.rs
@@ -4,7 +4,7 @@
 use crate::core::arch::{GpRegister, HartState, TrapReason};
 use crate::core::control_data::ConfidentialHart;
 use crate::core::memory_protector::HypervisorMemoryProtector;
-use crate::core::memory_tracker::{Allocated, Page, UnAllocated};
+use crate::core::page_allocator::{Allocated, Page, UnAllocated};
 use crate::core::transformations::{
     ConvertToConfidentialVm, ExposeToHypervisor, GuestLoadPageFaultRequest, GuestLoadPageFaultResult, InterruptRequest,
     MmioLoadRequest, MmioStoreRequest, OpensbiRequest, ResumeRequest, SbiRequest, SbiResult, SbiVmRequest,
