+++ security-monitor/src/core/heap_allocator/mod.rs
@@ -2,13 +2,13 @@
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
 use crate::core::memory_layout::ConfidentialMemoryAddress;
-use allocator::MemoryAllocator;
+use allocator::HeapAllocator;
 
 mod allocator;
 
 /// global allocator allocates memory on the security monitor's heap.
 #[global_allocator]
-static mut HEAP_ALLOCATOR: MemoryAllocator = MemoryAllocator::empty();
+static mut HEAP_ALLOCATOR: HeapAllocator = HeapAllocator::empty();
 
 pub(super) fn init_heap(start_address: ConfidentialMemoryAddress, heap_size: usize) {
     debug!("Heap {:x}-{:x}", start_address.as_usize(), start_address.as_usize() + heap_size);
