+++ security-monitor/src/core/page_allocator/mod.rs
@@ -1,10 +1,10 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
-pub use memory_tracker::MemoryTracker;
 pub use page::{Allocated, Page, PageState, UnAllocated};
+pub use page_allocator::MemoryTracker;
 pub use shared_page::SharedPage;
 
-mod memory_tracker;
 mod page;
+mod page_allocator;
 mod shared_page;
